Do reverse loops run faster than forward loops?
A blog post [1] inspected the byte code of simple "forward" and
"reverse" loops, found the latter to produce one less instruction, and
concluded it must be faster. Unfortunately the post did not include a
benchmark to support that conclusion. Let's find out!

A "forward" loop is a typical loop:

    for (int i = 0; i < 42; i++) {
        thing(i);
    }

A "reverse" loop is the semantically equivalent version that traverses
the opposite direction:

    for (int i = 42; i >= 0; i--) {
        thing(i);
    }

This transformation requires the operation to be associative, which
certainly is no guarantee.

For each forward and reverse loop there is a benchmark that caches the
loop termination condition ("Expression" in the specification [2]) and
one that doesn't, on the grounds that this saves another instruction.
There are two sets of these four benchmarks, one that uses only the loop
iterator and one that addresses the corresponding array element, on the
grounds that that's why you'd have an array to begin with.

[1] https://medium.com/@TravCav/why-reverse-loops-are-faster-a09d65473006
[2] https://docs.oracle.com/javase/specs/jls/se7/html/jls-14.html#jls-14.14